api/requirements.txt
# API server — install: pip install -r requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
anthropic>=0.45.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
joblib>=1.3.0
nltk>=3.8.0
pymongo>=4.6.0
